+++ internal/sdkprovider/service/pg/pg.go
@@ -12,10 +12,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil"
-	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig"
-	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/apiconvert"
-	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/dist"
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/stateupgrader"
+	"github.com/aiven/terraform-provider-aiven/internal/sdkprovider/userconfig/service"
 )
 
 func aivenPGSchema() map[string]*schema.Schema {
@@ -80,7 +78,7 @@ func aivenPGSchema() map[string]*schema.Schema {
 			},
 		},
 	}
-	schemaPG[schemautil.ServiceTypePG+"_user_config"] = dist.ServiceTypePg()
+	schemaPG[schemautil.ServiceTypePG+"_user_config"] = service.GetUserConfig(schemautil.ServiceTypePG)
 
 	return schemaPG
 }
@@ -135,11 +133,10 @@ func resourceServicePGUpdate(ctx context.Context, d *schema.ResourceData, m inte
 		return diag.FromErr(err)
 	}
 
-	userConfig, err := apiconvert.ToAPI(userconfig.ServiceTypes, "pg", d)
+	userConfig, err := schemautil.ExpandService(schemautil.ServiceTypePG, d)
 	if err != nil {
 		return diag.FromErr(err)
 	}
-
 	if userConfig["pg_version"] != nil {
 		s, err := client.Services.Get(ctx, projectName, serviceName)
 		if err != nil {