+++ CHANGELOG.md
@@ -14,6 +14,9 @@ nav_order: 1
 <!-- TODO: uncomment when dragonfly is supported -->
 <!-- - Dragonfly support -->
 - Fix `aiven_kafka_schema` JSON object diff suppress function
+- Add new user config generator
+- Use `TypeSet` for `ip_filter`, `ip_filter_string` fields
+
 
 ## [4.13.3] - 2024-01-29